.github/scripts/requirements.txt
google-genai>=0.3.0
openai>=1.54.0  # Optional: fallback if you want to support both
